pyproject.toml
[tool.poetry]
name = "django-starter"
version = "0.1.0"
description = "Django tutorial"
authors = ["Johannes Schriewer <j.schriewer@anfe.ma>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "<5"
lorem-text = "^2.1"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2.0"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
